+++ test/init.test.js
@@ -17,11 +17,39 @@ const exec = promisify(execCallback);
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const tempDir = path.join(__dirname, "temp");
 
-function verifyPackageName(expectedName) {
-  const packageJsonPath = path.join(tempDir, "package.json");
+function readPackageJson() {
+  const packageJsonPath = path.join(tempDir, 'package.json');
   const packageJsonContent = readFileSync(packageJsonPath, "utf8");
-  const packageJson = JSON.parse(packageJsonContent);
+  return JSON.parse(packageJsonContent);
+}
 
+function hasNodemon() {
+  const packageJson = readPackageJson();
+  if (!packageJson.scripts) {
+    return false;
+  }
+
+  if (!packageJson.scripts.dev) {
+    return false;
+  }
+
+  if (!packageJson.scripts.dev.includes('nodemon')) {
+    return false;
+  }
+
+  if (!packageJson.devDependencies) {
+    return false;
+  }
+
+  if (!packageJson.devDependencies.nodemon) {
+    return false;
+  }
+
+  return true;
+}
+
+function verifyPackageName(expectedName) {
+  const packageJson = readPackageJson();
   expect(packageJson.name).toBe(expectedName);
 }
 
@@ -89,25 +117,29 @@ describe("init", () => {
     clearTempDirectory();
   });
 
-  test("no templates passed, should default to basic", async () => {
+  test("no templates passed, should default to basic with nodemon", async () => {
     const originalHash = computeSHA256Hash(
       path.join(__dirname, "..", "templates", "basic")
     );
     await exec(`node ../../bin/index.js init`, { cwd: tempDir });
     const commandHash = computeSHA256Hash(tempDir);
     expect(commandHash).toEqual(originalHash);
+
+    expect(hasNodemon()).toBe(true);
   });
 
-  test("basic", async () => {
+  test("basic with nodemon", async () => {
     const originalHash = computeSHA256Hash(
       path.join(__dirname, "..", "templates", "basic")
     );
     await exec(`node ../../bin/index.js init -t basic`, { cwd: tempDir });
     const commandHash = computeSHA256Hash(tempDir);
     expect(commandHash).toEqual(originalHash);
+
+    expect(hasNodemon()).toBe(true);
   });
 
-  test("express_pg_sequelize", async () => {
+  test("express_pg_sequelize with nodemon", async () => {
     const originalHash = computeSHA256Hash(
       path.join(__dirname, "..", "templates", "express_pg_sequelize")
     );
@@ -116,9 +148,11 @@ describe("init", () => {
     });
     const commandHash = computeSHA256Hash(tempDir);
     expect(commandHash).toEqual(originalHash);
+
+    expect(hasNodemon()).toBe(true);
   }, 10000);
 
-  test("express_mysql", async () => {
+  test("express_mysql with nodemon", async () => {
     const originalHash = computeSHA256Hash(
       path.join(__dirname, "..", "templates", "express_mysql")
     );
@@ -127,9 +161,11 @@ describe("init", () => {
     });
     const commandHash = computeSHA256Hash(tempDir);
     expect(commandHash).toEqual(originalHash);
+
+    expect(hasNodemon()).toBe(true);
   }, 10000);
 
-  test("express_oauth_microsoft", async () => {
+  test("express_oauth_microsoft with nodemon", async () => {
     const originalHash = computeSHA256Hash(
       path.join(__dirname, "..", "templates", "express_oauth_microsoft")
     );
@@ -138,6 +174,8 @@ describe("init", () => {
     });
     const commandHash = computeSHA256Hash(tempDir);
     expect(commandHash).toEqual(originalHash);
+
+    expect(hasNodemon()).toBe(true);
   }, 10000);
 
   test("invalid template name passed", async () => {
@@ -204,4 +242,65 @@ describe("init", () => {
     });
     verifyPackageName(validName);
   });
+
+  // Nodemon flag tests.
+  test('no template passed, should default to basic template without nodemon', async () => {
+    await exec('node ../../bin/index.js init --remove-nodemon', { cwd: tempDir });
+    const packageJson = readPackageJson();
+
+    expect(packageJson.scripts.start).not.toContain('nodemon');
+    expect(packageJson.scripts.dev).toBeUndefined();
+
+    if (packageJson.devDependencies) {
+      expect(packageJson.devDependencies).not.toHaveProperty('nodemon');
+    }
+  }, 10000);
+
+  test('basic without nodemon', async () => {
+    await exec('node ../../bin/index.js init -t basic --remove-nodemon', { cwd: tempDir });
+    const packageJson = readPackageJson();
+
+    expect(packageJson.scripts.start).not.toContain('nodemon');
+    expect(packageJson.scripts.dev).toBeUndefined();
+
+    if (packageJson.devDependencies) {
+      expect(packageJson.devDependencies).not.toHaveProperty('nodemon');
+    }
+  }, 10000);
+
+  test('express_pg_sequelize without nodemon', async () => {
+    await exec('node ../../bin/index.js init -t express_pg_sequelize --remove-nodemon', { cwd: tempDir, });
+    const packageJson = readPackageJson();
+
+    expect(packageJson.scripts.start).not.toContain('nodemon');
+    expect(packageJson.scripts.dev).toBeUndefined();
+
+    if (packageJson.devDependencies) {
+      expect(packageJson.devDependencies).not.toHaveProperty('nodemon');
+    }
+  }, 10000);
+
+  test('express_mysql without nodemon', async () => {
+    await exec('node ../../bin/index.js init -t express_mysql --remove-nodemon', { cwd: tempDir, });
+    const packageJson = readPackageJson();
+
+    expect(packageJson.scripts.start).not.toContain('nodemon');
+    expect(packageJson.scripts.dev).toBeUndefined();
+
+    if (packageJson.devDependencies) {
+      expect(packageJson.devDependencies).not.toHaveProperty('nodemon');
+    }
+  }, 10000);
+
+  test('express_oauth_microsoft without nodemon', async () => {
+    await exec('node ../../bin/index.js init -t express_oauth_microsoft --remove-nodemon', { cwd: tempDir, });
+    const packageJson = readPackageJson();
+
+    expect(packageJson.scripts.start).not.toContain('nodemon');
+    expect(packageJson.scripts.dev).toBeUndefined();
+
+    if (packageJson.devDependencies) {
+      expect(packageJson.devDependencies).not.toHaveProperty('nodemon');
+    }
+  }, 10000);
 });
